requirements.txt
Django==1.8.5
django-allauth==0.23.0
django-crispy-forms==1.5.2
django-filter==0.11.0
django-report-builder==3.2.0
django-report-utils==0.3.13
django-rest-auth==0.5.0
djangorestframework==3.3.0
jdcal==1.0
Markdown==2.6.3
oauthlib==1.0.3
openpyxl==2.2.1
pbr==1.8.1
PyJWT==1.4.0
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
stevedore==1.9.0
virtualenv==1.11.4
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wheel==0.24.0
